+++ wand/api.py
@@ -1413,7 +1413,11 @@ if platform.system() == 'Windows':
         libc = ctypes.CDLL(msvcrt)
 else:
     if platform.system() == 'Darwin':
-        libc = ctypes.cdll.LoadLibrary('libc.dylib')
+        try:
+            libc = ctypes.cdll.LoadLibrary('libc.dylib')
+        except OSError:
+            # In case of El Capitan SIP
+            libc = ctypes.cdll.LoadLibrary('/usr/lib/libc.dylib')
     elif platform.system() == 'FreeBSD':
         libc = ctypes.cdll.LoadLibrary(ctypes.util.find_library('c'))
     else: